==> 13.pop/compensation/2.filter_same_tr.R <==
#compensation in population level
library(data.table)

##### filter ##### 
ann=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/uORF_variant_ann.id",sep='\t',header=T)
info=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/2.vep_UTRannotator_new/stats/varid_type_sort_MAF.txt")
info=info[info$type1=="SNP"&info$type %in% c("uAUG_gained","uAUG_lost"),]
info=info[,c("id","stat","MAF","type","type_new")]
colnames(info)=c("id","stat","MAF","type_old","type_new")
ann_snp=merge(ann,info,by="id")
#ann_snp[ann_snp$type_old!=ann_snp$type] #check

extract_same_tr <- function(f,ann_snp){
  anngain=unique(ann_snp[type_new=="uAUG_gained",c("chr","pos","ref","alt","freq","stat","FBgn")])
  annloss=unique(ann_snp[type_new=="uAUG_lost",c("chr","pos","ref","alt","freq","stat","FBgn")])
  
  colnames(anngain)=c("gain_chr","gain_pos","gain_ref","gain_alt","gain_freq","stat1","gain_FBgn")
  colnames(annloss)=c("loss_chr","loss_pos","loss_ref","loss_alt","loss_freq","stat2","loss_FBgn")
  
  f2=merge(f,anngain,by=c("gain_chr","gain_pos","gain_ref","gain_alt"),allow.cartesian=TRUE)
  f3=merge(f2,annloss,by=c("loss_chr","loss_pos","loss_ref","loss_alt"),allow.cartesian=TRUE)
  f4=f3[f3$gain_FBgn==f3$loss_FBgn,]
  
  anngain=unique(ann_snp[type_new=="uAUG_gained",c("chr","pos","ref","alt","FBgn","FBtr")])
  annloss=unique(ann_snp[type_new=="uAUG_lost",c("chr","pos","ref","alt","FBgn","FBtr")])
  
  colnames(anngain)=c("gain_chr","gain_pos","gain_ref","gain_alt","gain_FBgn","gain_FBtr")
  colnames(annloss)=c("loss_chr","loss_pos","loss_ref","loss_alt","loss_FBgn","loss_FBtr")
  f2=merge(f4,anngain,by=c("gain_chr","gain_pos","gain_ref","gain_alt","gain_FBgn"),allow.cartesian=TRUE)
  f3=merge(f2,annloss,by=c("loss_chr","loss_pos","loss_ref","loss_alt","loss_FBgn"),allow.cartesian=TRUE)
  f4_out=f3[f3$gain_FBtr==f3$loss_FBtr,]
  
  return(f4_out)
}

#CN
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/CN_gainloss_correlation_homo_2L.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_2L=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/CN_gainloss_correlation_homo_2R.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_2R=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/CN_gainloss_correlation_homo_3L.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_3L=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/CN_gainloss_correlation_homo_3R.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_3R=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/CN_gainloss_correlation_homo_4.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f$gain_chr=as.character(f$gain_chr)
f$loss_chr=as.character(f$loss_chr)
f_4=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/CN_gainloss_correlation_homo_X.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_X=extract_same_tr(f,ann_snp)
CN=rbind(f_2L,f_2R,f_3L,f_3R,f_4,f_X)

#CN$q=p.adjust(CN$p,method = "BH")
#CN$logq=-log10(CN$q)
#CN=CN[,c("gain_chr","gain_pos","gain_ref","gain_alt","gain_freq","stat1","loss_chr","loss_pos","loss_ref","loss_alt","loss_freq","stat2","gain_FBgn","gain_FBtr",
#       "AB","Ab","aB","ab","unknown","odd_ratio","p","q","logq")]
#fwrite(CN,"/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/CN_gain_loss_cor_filter.txt",sep='\t',na="NA",quote=F)
CN=CN[,c("gain_chr","gain_pos","gain_ref","gain_alt","gain_freq","stat1","loss_chr","loss_pos","loss_ref","loss_alt","loss_freq","stat2","gain_FBgn","gain_FBtr",
          "AB","Ab","aB","ab")]
CN$AB_new=CN$AB
CN$Ab_new=CN$Ab
CN$aB_new=CN$aB
CN$ab_new=CN$ab
CN[CN$stat1=="der"&CN$stat2=="anc",]$AB_new=CN[CN$stat1=="der"&CN$stat2=="anc",]$aB
CN[CN$stat1=="der"&CN$stat2=="anc",]$aB_new=CN[CN$stat1=="der"&CN$stat2=="anc",]$AB
CN[CN$stat1=="der"&CN$stat2=="anc",]$Ab_new=CN[CN$stat1=="der"&CN$stat2=="anc",]$ab
CN[CN$stat1=="der"&CN$stat2=="anc",]$ab_new=CN[CN$stat1=="der"&CN$stat2=="anc",]$Ab
CN[CN$stat1=="anc"&CN$stat2=="der",]$AB_new=CN[CN$stat1=="anc"&CN$stat2=="der",]$Ab
CN[CN$stat1=="anc"&CN$stat2=="der",]$aB_new=CN[CN$stat1=="anc"&CN$stat2=="der",]$ab
CN[CN$stat1=="anc"&CN$stat2=="der",]$Ab_new=CN[CN$stat1=="anc"&CN$stat2=="der",]$AB
CN[CN$stat1=="anc"&CN$stat2=="der",]$ab_new=CN[CN$stat1=="anc"&CN$stat2=="der",]$aB
CN[CN$stat1=="der"&CN$stat2=="der",]$AB_new=CN[CN$stat1=="der"&CN$stat2=="der",]$ab
CN[CN$stat1=="der"&CN$stat2=="der",]$aB_new=CN[CN$stat1=="der"&CN$stat2=="der",]$Ab
CN[CN$stat1=="der"&CN$stat2=="der",]$Ab_new=CN[CN$stat1=="der"&CN$stat2=="der",]$aB
CN[CN$stat1=="der"&CN$stat2=="der",]$ab_new=CN[CN$stat1=="der"&CN$stat2=="der",]$AB
fwrite(CN,"/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/CN_gain_loss_filter_new.txt",sep='\t',na="NA",quote=F)

#FR
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/FR_gainloss_correlation_homo_2L.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_2L=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/FR_gainloss_correlation_homo_2R.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_2R=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/FR_gainloss_correlation_homo_3L.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_3L=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/FR_gainloss_correlation_homo_3R.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_3R=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/FR_gainloss_correlation_homo_4.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f$gain_chr=as.character(f$gain_chr)
f$loss_chr=as.character(f$loss_chr)
f_4=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/FR_gainloss_correlation_homo_X.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_X=extract_same_tr(f,ann_snp)
FR=rbind(f_2L,f_2R,f_3L,f_3R,f_4,f_X)

FR=FR[,c("gain_chr","gain_pos","gain_ref","gain_alt","gain_freq","stat1","loss_chr","loss_pos","loss_ref","loss_alt","loss_freq","stat2","gain_FBgn","gain_FBtr",
         "AB","Ab","aB","ab")]
FR$AB_new=FR$AB
FR$Ab_new=FR$Ab
FR$aB_new=FR$aB
FR$ab_new=FR$ab
FR[FR$stat1=="der"&FR$stat2=="anc",]$AB_new=FR[FR$stat1=="der"&FR$stat2=="anc",]$aB
FR[FR$stat1=="der"&FR$stat2=="anc",]$aB_new=FR[FR$stat1=="der"&FR$stat2=="anc",]$AB
FR[FR$stat1=="der"&FR$stat2=="anc",]$Ab_new=FR[FR$stat1=="der"&FR$stat2=="anc",]$ab
FR[FR$stat1=="der"&FR$stat2=="anc",]$ab_new=FR[FR$stat1=="der"&FR$stat2=="anc",]$Ab
FR[FR$stat1=="anc"&FR$stat2=="der",]$AB_new=FR[FR$stat1=="anc"&FR$stat2=="der",]$Ab
FR[FR$stat1=="anc"&FR$stat2=="der",]$aB_new=FR[FR$stat1=="anc"&FR$stat2=="der",]$ab
FR[FR$stat1=="anc"&FR$stat2=="der",]$Ab_new=FR[FR$stat1=="anc"&FR$stat2=="der",]$AB
FR[FR$stat1=="anc"&FR$stat2=="der",]$ab_new=FR[FR$stat1=="anc"&FR$stat2=="der",]$aB
FR[FR$stat1=="der"&FR$stat2=="der",]$AB_new=FR[FR$stat1=="der"&FR$stat2=="der",]$ab
FR[FR$stat1=="der"&FR$stat2=="der",]$aB_new=FR[FR$stat1=="der"&FR$stat2=="der",]$Ab
FR[FR$stat1=="der"&FR$stat2=="der",]$Ab_new=FR[FR$stat1=="der"&FR$stat2=="der",]$aB
FR[FR$stat1=="der"&FR$stat2=="der",]$ab_new=FR[FR$stat1=="der"&FR$stat2=="der",]$AB
fwrite(FR,"/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/FR_gain_loss_filter_new.txt",sep='\t',na="NA",quote=F)

#RAL
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/RAL_gainloss_correlation_homo_2L.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_2L=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/RAL_gainloss_correlation_homo_2R.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_2R=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/RAL_gainloss_correlation_homo_3L.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_3L=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/RAL_gainloss_correlation_homo_3R.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_3R=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/RAL_gainloss_correlation_homo_4.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f$gain_chr=as.character(f$gain_chr)
f$loss_chr=as.character(f$loss_chr)
f_4=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/RAL_gainloss_correlation_homo_X.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_X=extract_same_tr(f,ann_snp)
RAL=rbind(f_2L,f_2R,f_3L,f_3R,f_4,f_X)

RAL=RAL[,c("gain_chr","gain_pos","gain_ref","gain_alt","gain_freq","stat1","loss_chr","loss_pos","loss_ref","loss_alt","loss_freq","stat2","gain_FBgn","gain_FBtr",
         "AB","Ab","aB","ab")]
RAL$AB_new=RAL$AB
RAL$Ab_new=RAL$Ab
RAL$aB_new=RAL$aB
RAL$ab_new=RAL$ab
RAL[RAL$stat1=="der"&RAL$stat2=="anc",]$AB_new=RAL[RAL$stat1=="der"&RAL$stat2=="anc",]$aB
RAL[RAL$stat1=="der"&RAL$stat2=="anc",]$aB_new=RAL[RAL$stat1=="der"&RAL$stat2=="anc",]$AB
RAL[RAL$stat1=="der"&RAL$stat2=="anc",]$Ab_new=RAL[RAL$stat1=="der"&RAL$stat2=="anc",]$ab
RAL[RAL$stat1=="der"&RAL$stat2=="anc",]$ab_new=RAL[RAL$stat1=="der"&RAL$stat2=="anc",]$Ab
RAL[RAL$stat1=="anc"&RAL$stat2=="der",]$AB_new=RAL[RAL$stat1=="anc"&RAL$stat2=="der",]$Ab
RAL[RAL$stat1=="anc"&RAL$stat2=="der",]$aB_new=RAL[RAL$stat1=="anc"&RAL$stat2=="der",]$ab
RAL[RAL$stat1=="anc"&RAL$stat2=="der",]$Ab_new=RAL[RAL$stat1=="anc"&RAL$stat2=="der",]$AB
RAL[RAL$stat1=="anc"&RAL$stat2=="der",]$ab_new=RAL[RAL$stat1=="anc"&RAL$stat2=="der",]$aB
RAL[RAL$stat1=="der"&RAL$stat2=="der",]$AB_new=RAL[RAL$stat1=="der"&RAL$stat2=="der",]$ab
RAL[RAL$stat1=="der"&RAL$stat2=="der",]$aB_new=RAL[RAL$stat1=="der"&RAL$stat2=="der",]$Ab
RAL[RAL$stat1=="der"&RAL$stat2=="der",]$Ab_new=RAL[RAL$stat1=="der"&RAL$stat2=="der",]$aB
RAL[RAL$stat1=="der"&RAL$stat2=="der",]$ab_new=RAL[RAL$stat1=="der"&RAL$stat2=="der",]$AB
fwrite(RAL,"/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/RAL_gain_loss_filter_new.txt",sep='\t',na="NA",quote=F)


#EF
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/EF_gainloss_correlation_homo_2L.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_2L=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/EF_gainloss_correlation_homo_2R.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_2R=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/EF_gainloss_correlation_homo_3L.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_3L=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/EF_gainloss_correlation_homo_3R.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_3R=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/EF_gainloss_correlation_homo_4.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f$gain_chr=as.character(f$gain_chr)
f$loss_chr=as.character(f$loss_chr)
f_4=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/EF_gainloss_correlation_homo_X.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_X=extract_same_tr(f,ann_snp)
EF=rbind(f_2L,f_2R,f_3L,f_3R,f_4,f_X)

EF=EF[,c("gain_chr","gain_pos","gain_ref","gain_alt","gain_freq","stat1","loss_chr","loss_pos","loss_ref","loss_alt","loss_freq","stat2","gain_FBgn","gain_FBtr",
           "AB","Ab","aB","ab")]
EF$AB_new=EF$AB
EF$Ab_new=EF$Ab
EF$aB_new=EF$aB
EF$ab_new=EF$ab
EF[EF$stat1=="der"&EF$stat2=="anc",]$AB_new=EF[EF$stat1=="der"&EF$stat2=="anc",]$aB
EF[EF$stat1=="der"&EF$stat2=="anc",]$aB_new=EF[EF$stat1=="der"&EF$stat2=="anc",]$AB
EF[EF$stat1=="der"&EF$stat2=="anc",]$Ab_new=EF[EF$stat1=="der"&EF$stat2=="anc",]$ab
EF[EF$stat1=="der"&EF$stat2=="anc",]$ab_new=EF[EF$stat1=="der"&EF$stat2=="anc",]$Ab
EF[EF$stat1=="anc"&EF$stat2=="der",]$AB_new=EF[EF$stat1=="anc"&EF$stat2=="der",]$Ab
EF[EF$stat1=="anc"&EF$stat2=="der",]$aB_new=EF[EF$stat1=="anc"&EF$stat2=="der",]$ab
EF[EF$stat1=="anc"&EF$stat2=="der",]$Ab_new=EF[EF$stat1=="anc"&EF$stat2=="der",]$AB
EF[EF$stat1=="anc"&EF$stat2=="der",]$ab_new=EF[EF$stat1=="anc"&EF$stat2=="der",]$aB
EF[EF$stat1=="der"&EF$stat2=="der",]$AB_new=EF[EF$stat1=="der"&EF$stat2=="der",]$ab
EF[EF$stat1=="der"&EF$stat2=="der",]$aB_new=EF[EF$stat1=="der"&EF$stat2=="der",]$Ab
EF[EF$stat1=="der"&EF$stat2=="der",]$Ab_new=EF[EF$stat1=="der"&EF$stat2=="der",]$aB
EF[EF$stat1=="der"&EF$stat2=="der",]$ab_new=EF[EF$stat1=="der"&EF$stat2=="der",]$AB
fwrite(EF,"/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/EF_gain_loss_filter_new.txt",sep='\t',na="NA",quote=F)

#SD
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/SD_gainloss_correlation_homo_2L.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_2L=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/SD_gainloss_correlation_homo_2R.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_2R=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/SD_gainloss_correlation_homo_3L.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_3L=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/SD_gainloss_correlation_homo_3R.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_3R=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/SD_gainloss_correlation_homo_4.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f$gain_chr=as.character(f$gain_chr)
f$loss_chr=as.character(f$loss_chr)
f_4=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/SD_gainloss_correlation_homo_X.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_X=extract_same_tr(f,ann_snp)
SD=rbind(f_2L,f_2R,f_3L,f_3R,f_4,f_X)

SD=SD[,c("gain_chr","gain_pos","gain_ref","gain_alt","gain_freq","stat1","loss_chr","loss_pos","loss_ref","loss_alt","loss_freq","stat2","gain_FBgn","gain_FBtr",
         "AB","Ab","aB","ab")]
SD$AB_new=SD$AB
SD$Ab_new=SD$Ab
SD$aB_new=SD$aB
SD$ab_new=SD$ab
SD[SD$stat1=="der"&SD$stat2=="anc",]$AB_new=SD[SD$stat1=="der"&SD$stat2=="anc",]$aB
SD[SD$stat1=="der"&SD$stat2=="anc",]$aB_new=SD[SD$stat1=="der"&SD$stat2=="anc",]$AB
SD[SD$stat1=="der"&SD$stat2=="anc",]$Ab_new=SD[SD$stat1=="der"&SD$stat2=="anc",]$ab
SD[SD$stat1=="der"&SD$stat2=="anc",]$ab_new=SD[SD$stat1=="der"&SD$stat2=="anc",]$Ab
SD[SD$stat1=="anc"&SD$stat2=="der",]$AB_new=SD[SD$stat1=="anc"&SD$stat2=="der",]$Ab
SD[SD$stat1=="anc"&SD$stat2=="der",]$aB_new=SD[SD$stat1=="anc"&SD$stat2=="der",]$ab
SD[SD$stat1=="anc"&SD$stat2=="der",]$Ab_new=SD[SD$stat1=="anc"&SD$stat2=="der",]$AB
SD[SD$stat1=="anc"&SD$stat2=="der",]$ab_new=SD[SD$stat1=="anc"&SD$stat2=="der",]$aB
SD[SD$stat1=="der"&SD$stat2=="der",]$AB_new=SD[SD$stat1=="der"&SD$stat2=="der",]$ab
SD[SD$stat1=="der"&SD$stat2=="der",]$aB_new=SD[SD$stat1=="der"&SD$stat2=="der",]$Ab
SD[SD$stat1=="der"&SD$stat2=="der",]$Ab_new=SD[SD$stat1=="der"&SD$stat2=="der",]$aB
SD[SD$stat1=="der"&SD$stat2=="der",]$ab_new=SD[SD$stat1=="der"&SD$stat2=="der",]$AB
fwrite(SD,"/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/SD_gain_loss_filter_new.txt",sep='\t',na="NA",quote=F)

#ZI
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/ZI_gainloss_correlation_homo_2L.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_2L=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/ZI_gainloss_correlation_homo_2R.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_2R=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/ZI_gainloss_correlation_homo_3L.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_3L=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/ZI_gainloss_correlation_homo_3R.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_3R=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/ZI_gainloss_correlation_homo_4.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f$gain_chr=as.character(f$gain_chr)
f$loss_chr=as.character(f$loss_chr)
f_4=extract_same_tr(f,ann_snp)
f=fread("/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/ZI_gainloss_correlation_homo_X.txt",header=T)
colnames(f)=c("gain_chr","gain_pos","gain_ref","gain_alt","loss_chr","loss_pos","loss_ref","loss_alt","AB","Ab","aB","ab","unknown","odd_ratio","p")
f_X=extract_same_tr(f,ann_snp)
ZI=rbind(f_2L,f_2R,f_3L,f_3R,f_4,f_X)

ZI=ZI[,c("gain_chr","gain_pos","gain_ref","gain_alt","gain_freq","stat1","loss_chr","loss_pos","loss_ref","loss_alt","loss_freq","stat2","gain_FBgn","gain_FBtr",
         "AB","Ab","aB","ab")]
ZI$AB_new=ZI$AB
ZI$Ab_new=ZI$Ab
ZI$aB_new=ZI$aB
ZI$ab_new=ZI$ab
ZI[ZI$stat1=="der"&ZI$stat2=="anc",]$AB_new=ZI[ZI$stat1=="der"&ZI$stat2=="anc",]$aB
ZI[ZI$stat1=="der"&ZI$stat2=="anc",]$aB_new=ZI[ZI$stat1=="der"&ZI$stat2=="anc",]$AB
ZI[ZI$stat1=="der"&ZI$stat2=="anc",]$Ab_new=ZI[ZI$stat1=="der"&ZI$stat2=="anc",]$ab
ZI[ZI$stat1=="der"&ZI$stat2=="anc",]$ab_new=ZI[ZI$stat1=="der"&ZI$stat2=="anc",]$Ab
ZI[ZI$stat1=="anc"&ZI$stat2=="der",]$AB_new=ZI[ZI$stat1=="anc"&ZI$stat2=="der",]$Ab
ZI[ZI$stat1=="anc"&ZI$stat2=="der",]$aB_new=ZI[ZI$stat1=="anc"&ZI$stat2=="der",]$ab
ZI[ZI$stat1=="anc"&ZI$stat2=="der",]$Ab_new=ZI[ZI$stat1=="anc"&ZI$stat2=="der",]$AB
ZI[ZI$stat1=="anc"&ZI$stat2=="der",]$ab_new=ZI[ZI$stat1=="anc"&ZI$stat2=="der",]$aB
ZI[ZI$stat1=="der"&ZI$stat2=="der",]$AB_new=ZI[ZI$stat1=="der"&ZI$stat2=="der",]$ab
ZI[ZI$stat1=="der"&ZI$stat2=="der",]$aB_new=ZI[ZI$stat1=="der"&ZI$stat2=="der",]$Ab
ZI[ZI$stat1=="der"&ZI$stat2=="der",]$Ab_new=ZI[ZI$stat1=="der"&ZI$stat2=="der",]$aB
ZI[ZI$stat1=="der"&ZI$stat2=="der",]$ab_new=ZI[ZI$stat1=="der"&ZI$stat2=="der",]$AB
fwrite(ZI,"/gpfs2/liucl/lcl/analysis/uORF_gain_loss/pop_var/uORF/6.compensation/6pop/ZI_gain_loss_filter_new.txt",sep='\t',na="NA",quote=F)
